+++ pimms/longrange_utils.py
@@ -98,7 +98,7 @@ def build_LR_envelope_pairs(positions, LR_binary_array, type_grid, dimensions):
     if len(dimensions) == 2:
 
         for i in range(0, len(positions)):
-            (LR_tmp, SLR_tmp)  = inner_loops.extract_LR_pairs_from_position_2D(np.array(positions[i], dtype=int), LR_binary_array[i], type_grid, dimensions[0], dimensions[1])
+            (LR_tmp, SLR_tmp)  = inner_loops.extract_LR_pairs_from_position_2D(np.array(positions[i], dtype=np.int16), LR_binary_array[i], type_grid, dimensions[0], dimensions[1])
             
             if len(LR_tmp) > 0:
                 LR_list.append(LR_tmp)
@@ -176,7 +176,7 @@ def build_LR_envelope_pairs(positions, LR_binary_array, type_grid, dimensions):
     else:
 
         for i in range(0, len(positions)):
-            (LR_tmp, SLR_tmp)  = inner_loops.extract_LR_pairs_from_position_3D(np.array(positions[i], dtype=int), LR_binary_array[i], type_grid, dimensions[0], dimensions[1], dimensions[2])
+            (LR_tmp, SLR_tmp)  = inner_loops.extract_LR_pairs_from_position_3D(np.array(positions[i], dtype=np.int16), LR_binary_array[i], type_grid, dimensions[0], dimensions[1], dimensions[2])
 
             if len(LR_tmp) > 0:
                 LR_list.append(LR_tmp)
